refactor: remove stale functions that are no longer used and other cleanup
